# src/main/resources/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8085
  description: Generated server url
paths:
  /update/{id}:
    put:
      tags:
      - order-controller
      operationId: updateOrder
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              required:
              - file
              type: object
              properties:
                order:
                  $ref: '#/components/schemas/Order'
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Order'
...
  /upload:
    post:
      tags:
      - order-controller
      operationId: uploadFile
      requestBody:
        content:
          application/json:
            schema:
              required:
              - file
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/upload:
    post:
      tags:
      - file-controller
      operationId: uploadFile_1
      requestBody:
        content:
          application/json:
            schema:
              required:
              - file
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /addOrders:
    post:
      tags:
      - order-controller
      operationId: addOrders
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Order'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
  /addOrder:
    post:
      tags:
      - order-controller
      operationId: addOrder
      requestBody:
        content:
          application/json:
            schema:
              required:
              - file
              type: object
              properties:
                order:
                  $ref: '#/components/schemas/Order'
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Order'
...
  /orderByName/{name}:
    get:
      tags:
      - order-controller
      operationId: findIOrderByName
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Order'
  /orderById/{id}:
    get:
      tags:
      - order-controller
      operationId: findIOrderById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Order'
...
  /delete/{id}:
    delete:
      tags:
      - order-controller
      operationId: deleteOrder
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
components:
  schemas:
    Order:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        foodType:
          type: string
        price:
          type: number
          format: double
        filename:
          type: string
        data:
          type: array
          items:
            type: string
            format: byte
...